Move Ui class into bob.ui pkg
