+++ src/main/java/bob/Bob.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import bob.command.Command;
 import bob.command.WrongCommandException;
 import bob.parser.Parser;
+import bob.ui.Ui;
 
 public class Bob {
     public static void main(String[] args) {
