+++ src/main/java/bob/command/Command.java
@@ -1,6 +1,6 @@
 package bob.command;
 
-import bob.Ui;
+import bob.ui.Ui;
 
 public abstract class Command {
     public abstract void execute(Ui ui);
