+++ src/main/java/bob/command/EventCommand.java
@@ -2,9 +2,9 @@ package bob.command;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-import bob.Ui;
 import bob.task.Event;
 import bob.task.TaskList;
+import bob.ui.Ui;
 
 public class EventCommand extends Command {
