+++ src/main/java/bob/ui/Ui.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.ui;
 import java.util.List;
 import java.util.Scanner;
